updated to starting application point (tests currently disabled while being fixed)
