--- GradeBook/Enums/GradeBookType.cs
@@ -1,11 +1,0 @@
-﻿namespace GradeBook.Enums
-{
-    public enum GradeBookType
-    {
-        ESNU,
-        OneToFour,
-        Ranked,
-        SixPoint,
-        Standard
-    }
-}
